actualizacion de jpa

+++ src/main/java/ar/org/centro8/java/curso/proyecto_final_poo/model/repository/ClienteRepository.java
@@ -1,17 +1,12 @@
 package ar.org.centro8.java.curso.proyecto_final_poo.model.repository;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
-import org.springframework.data.jpa.repository.JpaRepository;
 
 import ar.org.centro8.java.curso.proyecto_final_poo.model.entity.Cliente;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
